refactor: extract episode export logic into internal exporter component
